map_to_minimum_cell in dials.symmetry
Map experiments/reflections to minimum cell at the start of dials.symmetry.
Move staticmethod in dials.cosym to function in dials.symmetry, also move
associated test to test_symmetry.py.

Fixes #992
